+++ src/GammonX/GammonX.Server.Tests/Utils/BotUtils.cs
@@ -5,7 +5,7 @@ namespace GammonX.Server.Tests.Utils
     public static class BotUtils
     {
         private static readonly HttpClient WildBgClient = new() { BaseAddress = new Uri("http://localhost:8082/bot/wildbg/") };
-        private static readonly HttpClient MarsClient = new() { BaseAddress = new Uri("http://localhost:8083/bot/mars/") };
+        private static readonly MarsClient MarsClient = new(null, null) { BaseAddress = new Uri("http://localhost:8083/bot/mars/") };
 
         public static IBotService GetBotService(string botHint)
         {